🏗 folder :  character 이미지 경로 수정
