+++ src/assets/illust.ts
@@ -1,4 +1,22 @@
 export const logo = "/img/logo.svg"
+
+export const character = [
+    "/img/character/ch_0.svg",
+    "/img/character/ch_1.svg",
+    "/img/character/ch_2.svg",
+    "/img/character/ch_3.svg",
+    "/img/character/ch_4.svg",
+    "/img/character/ch_5.svg",
+  ];
+  export const character_like = [
+    "/img/character/like_0.svg",
+    "/img/character/like_1.svg",
+    "/img/character/like_2.svg",
+    "/img/character/like_3.svg",
+    "/img/character/like_4.svg",
+    "/img/character/like_5.svg",
+  ];
+  
 export const illust = {
     character : "/img/illust/character.svg",
     report : "/img/illust/library_report.svg",
