+++ src/components/kitchen/RecipeCard.tsx
@@ -4,7 +4,7 @@ import colors from "../../assets/styles";
 import Image from "next/image";
 import { selectToken } from "../../store/tokenSlice";
 import { useSelector } from "react-redux";
-import { character_like } from "../../assets/character";
+import { character_like } from "../../assets/illust";
 import { ic_heart } from "../../assets/icon";
 import { route } from "../../assets/route";
